Move closure expression checking
