+++ src/typechecker/expressions.rs
@@ -19,7 +19,6 @@ use crate::ast::typed::*;
 use crate::ast::{AstType, Expression, Param};
 use crate::error::{Diagnostic, Span};
 
-use super::closures::collect_captures;
 use super::monomorphize_inference::InferenceConflict;
 use super::monomorphize_types::concrete_name_matches_generic;
 use super::{BehaviorBound, FuncInfo, TypeChecker};
